requirements.txt
pymavlink
pyserial
pandas
matplotlib